requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
openai
faiss-cpu
numpy
python-multipart
python-dotenv
langchain
langchain-openai
sentence-transformers
google-generativeai
